- Changed tick column definition int, forgot to update that (somewhere along the process it was an unsigned int, and now it's just an int)
git-svn-id: http://eathena.googlecode.com/svn/branches/stable@3274 9c4de8d1-4af9-ca9b-867f-b47ac08f6c80

+++ sql-files/upgrade_svn3273.sql
@@ -2,7 +2,7 @@ CREATE TABLE `sc_data` (
   `account_id` int(11) unsigned NOT NULL,
   `char_id` int(11) unsigned NOT NULL,
   `type` smallint(11) unsigned NOT NULL,
-  `tick` bigint(11) unsigned NOT NULL,
+  `tick` int(11) NOT NULL,
   `val1` int(11) NOT NULL default '0',
   `val2` int(11) NOT NULL default '0',
   `val3` int(11) NOT NULL default '0',
